[LB-20] refactor: 디렉터리 구조 변경
