+++ src/main/java/com/lgcms/consulting/dto/request/lecture/RemoteLectureRequest.java
@@ -1,0 +1,4 @@
+package com.lgcms.consulting.dto.request.lecture;
+
+public class RemoteLectureRequest {
+}
