--- src/main/java/com/lgcms/consulting/remote/lesson/dto/RemoteLectureRequest.java
@@ -1,4 +1,0 @@
-package com.lgcms.consulting.remote.lesson.dto;
-
-public class RemoteLectureRequest {
-}
